+++ modules/portal-security/portal-security-sso-ntlm/src/main/java/com/liferay/portal/security/sso/ntlm/configuration/NtlmConfiguration.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.ntlm.module.configuration;
+package com.liferay.portal.security.sso.ntlm.configuration;
 
 import aQute.bnd.annotation.metatype.Meta;
 
@@ -31,7 +31,7 @@ import com.liferay.configuration.admin.ConfigurationAdmin;
  */
 @ConfigurationAdmin(category = "platform")
 @Meta.OCD(
-	id = "com.liferay.portal.security.sso.ntlm.module.configuration.NtlmConfiguration",
+	id = "com.liferay.portal.security.sso.ntlm.configuration.NtlmConfiguration",
 	localization = "content/Language", name = "%ntlm.configuration.name"
 )
 public interface NtlmConfiguration {
