+++ modules/portal-security/portal-security-sso-ntlm/src/main/java/com/liferay/portal/security/sso/ntlm/configuration/definition/NtlmCompanyServiceConfigurationPidMapping.java
@@ -12,11 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.ntlm.module.configuration.definition;
+package com.liferay.portal.security.sso.ntlm.configuration.definition;
 
 import com.liferay.portal.kernel.settings.definition.ConfigurationPidMapping;
 import com.liferay.portal.security.sso.ntlm.constants.NtlmConstants;
-import com.liferay.portal.security.sso.ntlm.module.configuration.NtlmConfiguration;
+import com.liferay.portal.security.sso.ntlm.configuration.NtlmConfiguration;
 
 import org.osgi.service.component.annotations.Component;
