+++ modules/portal-security/portal-security-sso-ntlm/src/main/resources/META-INF/resources/com.liferay.portal.settings.web/init.jsp
@@ -22,7 +22,7 @@ page import="com.liferay.portal.kernel.settings.CompanyServiceSettingsLocator" %
 page import="com.liferay.portal.kernel.settings.ParameterMapSettingsLocator" %><%@
 page import="com.liferay.portal.kernel.util.Validator" %><%@
 page import="com.liferay.portal.security.sso.ntlm.constants.NtlmConstants" %><%@
-page import="com.liferay.portal.security.sso.ntlm.module.configuration.NtlmConfiguration" %><%@
+page import="com.liferay.portal.security.sso.ntlm.configuration.NtlmConfiguration" %><%@
 page import="com.liferay.portal.util.Portal" %>
 
 <%@ page import="javax.portlet.ActionRequest" %>
